Silence Win warnings
TBR=bmeurer@chromium.org
BUG=

Review URL: https://codereview.chromium.org/194893002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@19789 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/types.h
@@ -444,7 +444,7 @@ struct ZoneTypeConfig {
   }
   static int lub_bitset(Type* type) {
     ASSERT(is_class(type) || is_constant(type));
-    return tagged_get<intptr_t>(as_tagged(type), 0);
+    return static_cast<int>(tagged_get<intptr_t>(as_tagged(type), 0));
   }
 };
